Fix test issue 5

+++ packages/meditrak-app/app/assessment/specificQuestions/FileQuestion.jsx
@@ -3,15 +3,15 @@
  * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
  */
 
-import React, { useState } from 'react';
-import { StyleSheet, View } from 'react-native';
+import React, {useState} from 'react';
+import {StyleSheet, View} from 'react-native';
 import PropTypes from 'prop-types';
 import DocumentPicker from 'react-native-document-picker';
 import RNFS from 'react-native-fs';
 
-import { DEFAULT_PADDING } from '../../globalStyles';
-import { Button, Text } from '../../widgets';
-import { getFilenameFromUri } from '../../utilities';
+import {DEFAULT_PADDING} from '../../globalStyles';
+import {Button, Text} from '../../widgets';
+import {getFilenameFromUri} from '../../utilities';
 
 /*
  * Limit the max file size to:
@@ -99,7 +99,7 @@ const unlink = async path => {
   }
 };
 
-export const FileQuestion = ({ answer, onChangeAnswer }) => {
+export const FileQuestion = ({answer, onChangeAnswer}) => {
   const filename = getFilenameFromUri(answer);
 
   const [error, setError] = useState(null);
@@ -120,11 +120,11 @@ export const FileQuestion = ({ answer, onChangeAnswer }) => {
       // If we previously picked a file (stored a copy in app documents) and now picked a new one, we need to delete the old copy
       await unlink(answer);
 
-      const { fileCopyUri } = filePickerResponse;
+      const {fileCopyUri} = filePickerResponse;
 
       // Check file size
       // Note: from DocumentPicker docs: android does not guarantee the name or size will be present or accurate on the file picker response, so we read the newly copied file in the app documents which is a normal file
-      const { size: newSizeInBytes } = await RNFS.stat(fileCopyUri);
+      const {size: newSizeInBytes} = await RNFS.stat(fileCopyUri);
       if (newSizeInBytes > MAX_FILE_SIZE_BYTES) {
         await unlink(fileCopyUri);
         setSizeInBytes(null);
@@ -140,10 +140,14 @@ export const FileQuestion = ({ answer, onChangeAnswer }) => {
       onChangeAnswer(fileCopyUri);
       setSizeInBytes(newSizeInBytes);
     } catch (e) {
-      await unlink(answer);
-      onChangeAnswer(null);
-      setSizeInBytes(null);
-      setError(`Error: ${e.message}`);
+      if (DocumentPicker.isCancel(e)) {
+        // User cancelled, do nothing
+      } else {
+        await unlink(answer);
+        onChangeAnswer(null);
+        setSizeInBytes(null);
+        setError(`Error: ${e.message}`);
+      }
     }
   };
